LLR_Point_Detection-main/requirements.txt
scipy==1.10.0
typeguard
menpo
menpofit
tensorflow==2.12
tensorflow-addons==0.20.0
pillow
scikit-image
scikit-learn
numpy==1.23.5
matplotlib
pandas
pingouin
seaborn
statsmodels
notebook
configparser
torch==2.3
torchaudio==2.3
torchvision==0.18
opencv-python
matplotlib
python_requires=">=3.11,<3.12"
